% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radiant.R
\name{xsanalytics}
\alias{xsanalytics}
\title{Launch Radiant in the default browser}
\usage{
xsanalytics(app = c("marketing", "quant", "base"))
}
\arguments{
\item{app}{Choose the app to run. Either "base", "quant", or "marketing". "marketing" is the default}
}
\description{
Launch Radiant in the default browser
}
\details{
See \url{http://vnijs.github.io/radiant} for documentation and tutorials
}
\examples{
if (interactive()) {
  radiant("base")
  radiant("quant")
  radiant("marketing")
}
}

